Refactor query, suburl - move outside of the folder
