+++ src/query.ts
@@ -17,22 +17,20 @@
  * @param queryParams query params in object form
  * @returns url with encoded params
  */
-export function query(url: string, queryParams: object) {
+export const query = (url: string) => (queryParams: Record<PropertyKey, any>) => {
   const query = Object.keys(queryParams)
     .filter(k => k && queryParams[k] !== undefined)
-    .map((k): [string, string] => {
-      if (Array.isArray(queryParams[k])) {
-        return [queryParams[k].join(','), k];
-      } else if (typeof queryParams[k] === 'object') {
-        return [JSON.stringify(queryParams[k]), k];
-      }
-
-      return [queryParams[k], k];
-    })
-    .map(([value, key]) => `${encodeURIComponent(key)}=${encodeURI(value)}`)
+    .map((k): [string, string] =>
+      Array.isArray(queryParams[k])
+        ? [k, queryParams[k].join(',')]
+      : typeof queryParams[k] === 'object'
+        ? [k, JSON.stringify(queryParams[k])]
+      : [k, queryParams[k]]
+    )
+    .map(([key, value]) => `${encodeURIComponent(key)}=${encodeURI(value)}`)
     .join('&');
 
   const prefix = (String(url).indexOf('?') > -1 ? '&' : '?');
 
   return url + (query.length > 0 ? prefix + query : '');
-}
+};
