--- src/url/index.ts
@@ -1,2 +1,0 @@
-export * from './query';
-export * from './suburl';
